Moved from cve-search to NVD
Since the api from cve-search only updates like once a week, we
are now instead using the api provided from nvd.nist.gov, i.e.
from the "National Institute of Standards" in the US. We hope this
provides for a more fun experience with more frequent updates.

Added a couple of QoL improvements to the UI of the Vulnerabilities
skin.

Streamlines the location variables such that they will be sourced
from the Variables.inc file for various skins.